Add Video Views in Student Info
