+++ rg_instructor_analytics_log_collector/admin.py
@@ -27,3 +27,5 @@ admin.site.register(models.LogTable, admin.ModelAdmin)
 admin.site.register(models.EnrollmentByDay, admin.ModelAdmin)
 admin.site.register(models.EnrollmentByUser, admin.ModelAdmin)
 admin.site.register(models.LastProcessedLog, LastProcessedLogAdmin)
+admin.site.register(models.VideoViewsByUser, admin.ModelAdmin)
+admin.site.register(models.VideoViewsByBlock, admin.ModelAdmin)
