+++ rg_instructor_analytics_log_collector/processors/processor.py
@@ -5,10 +5,8 @@ import logging
 import operator
 import time
 
-from django.db.models import Q
-
-from rg_instructor_analytics_log_collector.models import LogTable
-from rg_instructor_analytics_log_collector.processors.base_pipeline import EnrollmentPipeline
+from rg_instructor_analytics_log_collector.processors.enrollment_pipeline import EnrollmentPipeline
+from rg_instructor_analytics_log_collector.processors.video_views_pipeline import VideoViewsPipeline
 
 log = logging.getLogger(__name__)
 
@@ -19,7 +17,8 @@ class Processor(object):
     """
 
     available_pipelines = [
-        EnrollmentPipeline()
+        EnrollmentPipeline(),
+        VideoViewsPipeline(),
     ]
 
     def __init__(self, alias_list, sleep_time):
@@ -32,23 +31,6 @@ class Processor(object):
         super(Processor, self).__init__()
         self.sleep_time = sleep_time
         self.pipelinies = filter(lambda x: x.alias in alias_list, self.available_pipelines)
-
-    def _get_query_for_pipeline(self, pipeline):
-        """
-        Return list of the raw logs wit type, that suitable for the given pipeline.
-        """
-        type_request = None
-        for pipeline_type in pipeline.supported_types:
-            if type_request is None:
-                type_request = Q(message_type__contains=pipeline_type)
-            else:
-                type_request |= Q(message_type__contains=pipeline_type)
-
-        query = LogTable.objects.filter(type_request)
-        last_processed_log_date = pipeline.retrieve_last_date()
-        if last_processed_log_date:
-            query = query.filter(log_time__lt=last_processed_log_date)
-        return query.order_by('log_time').all()
 
     def _sort(self, ordering, records):
         """
@@ -73,18 +55,18 @@ class Processor(object):
         """
         while True:
             for pipeline in self.pipelinies:
-                records = self._get_query_for_pipeline(pipeline)
-                last_record = records.last()
+                
+                records = pipeline.get_query()
                 if not records:
                     continue
-                if pipeline.format:  # ? getattr / hasattr
-                    records = filter(None, [pipeline.format(m) for m in records])
-                if pipeline.ordered_fields:
-                    records = self._sort(pipeline.ordered_fields, records)
-                if pipeline.aggregate:
+                last_record = records.last()
+                # Format raw log to the internal format.
+                records = filter(None, [pipeline.format(m) for m in records])
+                records = self._sort(pipeline.ordered_fields, records)
+                if pipeline.alias == EnrollmentPipeline.alias:
                     records = pipeline.aggregate(records)
-                for m in records:
-                    db_context = pipeline.load_database_contex and pipeline.load_database_contex(m) or None
-                    pipeline.push_to_database(m, db_context)
+                for record in records:
+                    db_context = pipeline.load_database_contex(record)
+                    pipeline.push_to_database(record, db_context)
                 pipeline.update_last_processed_log(last_record)
             time.sleep(self.sleep_time)
